put all the protected route code in 1 place first before sorting
