+++ package.json
@@ -15,7 +15,7 @@
   },
   "scripts": {
     "start": "PORT=8000 react-scripts start",
-    "start:win" : " set PORT=8000 && react-scripts start",
+    "start:win": " set PORT=8000 && react-scripts start",
     "build": "react-scripts build",
     "test": "react-scripts test",
     "eject": "react-scripts eject"
