+++ src/auth.js
