+++ src/pages/UserDashboardPage.js
@@ -38,6 +38,26 @@ class UserDashboardPage extends Component {
         }
     }
 
+    // Private Route
+
+    const PrivateRoute = ({ component: Component, ...rest }) => {
+        return (
+        <Route
+            {...rest}
+            render={props =>
+            fakeAuth.isAuthenticated === true ? (
+                <Component {...props} />
+            ) : (
+                <Redirect
+                to={{ pathname: "/login", state: { from: props.location } }}
+                />
+            )
+            }
+        />
+        );
+    };
+
+
     getGifsByUser = async () => {
         try {
             const response = await api.getGifsByUser(this.state.username)
